Rework undistract-me.sh
It looks weird but it works!  It's also much more complex than it
should.  I'm commiting it like that so that it exists in the history,
but afterwards I'll just port it to zsh-only.  I mean, I'm on bash right
now and I'm not using it so why keep it "portable" and complicated?