Overhaul vimrc (a bit)
CtrlP  ->  fzf
I like CtrlP but it's not fast and in general is a bit of a pain working
with.  Right now I'm only removing the plugin but next commit I'm
removing any relevant code too.

No more thematic
It was a nice idea but ultimately not that useful.

interestingwords  ->  wordhl
This is another "plugin" of mine, far simpler and smaller.  Eventually
I'm going to have to remove all the little plugins from vimrc and move
them in separate files.