make conky great again
Scrap (most) of the config and rewrite it in the new Lua-like format.
I based it on Harmattan though not much of it is left.

Refactor any Lua code.  Where before we were executing and parsing the
result of `transmission-remote' we now use the JSONRPC API directly
through a very minimal Lua implementation.  Does it really matter?  Not
much, the previous solution was good enough.  But who cares.