Update awesome config
Lots of changes really, most based on awesome-copycats.  New theme
(based on multicolor), different widgets, some more keybinds.  The
biggest thing though, for me, is the MPRIS integration.  You can see
what is playing both on the wibar as well as a "conky" like thing on the
desktop itself, which also shows the current cover!