sway: replace wlsunset with gammastep
wlsunset has some warts unfortunatelly:
* it won't properly set the gamma when the display is turned off (as in,
  if I leave the screen turned off all night, when I come back in the
  morning is all red)
* you can't easily toggle it on and off

So I replace it with gammastep which is much much closer to redshift
including its SIGUSR1 toggle.  I almost literally copied redshift's
config over.

I also added a small script to turn that SIGUSR1 toggle into a swith
with SIGUSR1 and SIGUSR2, to use it with mpv.  It's not perfect: if I
open 2 mpv instances, when the first one exits it will turn the screen
red (if it's nighttime of course) but at least it's better than leaving
it blue when the second one and getting blinded.  I could make it count
how many times it's been turned off and on again but then if an mpv
instance crashes that will leave the counter confused and that's getting
too complicated.