sway: use foot as terminal emulator
I love st but I haven't really found a good way to manage it's
configuration.

foot is almost as lightweight as st and I potentially faster.  It's also
wayland native!  With a little hack, bells set the window as urgent
which none of the other wayland emulators do (I don't know if it's their
fault or sway's).