sway: fixes, change font, upgrade to 1.6
Sway 1.6 has urgency support so no more `swaymsg` hacks in `foot`.

I also found out why `swaynag` wouldn't quit `sway`: it tries to run
`swaymsg exit` with `$TERMINAL` which was set to `st`!  I was setting
`$TERM` to `foot` and not `$TERMINAL` like I should.  That's now fixed
and I also changed `swaynag` to run the command directly and not through
the terminal.